Create Group Facet
While Group and System has always been different, they have shared
a single registry class. Here we not only create the facet but also
split the registry.
