+++ pint/facets/group/__init__.py
@@ -1,0 +1,15 @@
+"""
+    pint.facets.group
+    ~~~~~~~~~~~~~~~~~
+
+    Adds pint the capability to group units.
+
+    :copyright: 2022 by Pint Authors, see AUTHORS for more details.
+    :license: BSD, see LICENSE for more details.
+"""
+
+from .definitions import GroupDefinition
+from .objects import Group
+from .registry import GroupRegistry
+
+__all__ = [GroupDefinition, Group, GroupRegistry]
